refactor(types): вынести классы вывода типов ОСени в подпакет autumn
AutumnAnnotations, AutumnBeanIndex, AutumnComponentInferencer и
AutumnMetaAnnotationResolver (+ их тесты) перенесены в
types.inferencer.autumn. Добавлен package-info.java с @NullMarked.
